refactor: restructure directories
