+++ aptos/api/aux-ts/examples/devnet-amm-limit-trader.ts
@@ -8,7 +8,7 @@ import { AptosAccount, AptosClient } from "aptos";
 import { assert } from "console";
 import { WebSocket } from "ws";
 import { AU, DU } from "../src";
-import { PoolClient } from "../src/amm/pool";
+import { PoolClient } from "../src/pool/client";
 import { AuxClient } from "../src/client";
 import { FakeCoin } from "../src/coin";
