+++ aptos/api/aux-ts/examples/devnet-amm-trader.ts
@@ -7,7 +7,7 @@
 import { AptosAccount, AptosClient } from "aptos";
 import axios from "axios";
 import { AU, DU } from "../src";
-import { PoolClient } from "../src/amm/pool";
+import { PoolClient } from "../src/pool/client";
 import { AuxClient } from "../src/client";
 import { FakeCoin } from "../src/coin";
