+++ aptos/api/aux-ts/scripts/sim/live.ts
@@ -4,8 +4,8 @@ import { AptosAccount, HexString, Types } from "aptos";
 import assert from "assert";
 import _ from "lodash";
 import { Market, Vault } from "../../src";
-import { PoolClient } from "../../src/amm/pool";
-import type { SwapExactInInput } from "../../src/amm/schema";
+import { PoolClient } from "../../src/pool/client";
+import type { SwapExactInInput } from "../../src/pool/schema";
 import { AuxClient } from "../../src/client";
 import { OrderType } from "../../src/clob/core/mutation";
 import { ALL_FAKE_COINS, ALL_FAKE_VOLATILES, FakeCoin } from "../../src/coin";
