+++ aptos/api/aux-ts/src/pool/schema.ts
