+++ aptos/api/aux-ts/test/amm.spec.ts
@@ -2,8 +2,8 @@ import { AptosAccount } from "aptos";
 import * as assert from "assert";
 import { describe, it } from "mocha";
 import { Vault } from "../src";
-import { PoolClient } from "../src/amm/pool";
-import type { ConstantProduct } from "../src/amm/schema";
+import { PoolClient } from "../src/pool/client";
+import type { ConstantProduct } from "../src/pool/schema";
 import { AuxClient } from "../src/client";
 import { FakeCoin } from "../src/coin";
 import { AuxEnv } from "../src/env";
