+++ aptos/api/aux-ts/test/pool.ts
@@ -1,6 +1,6 @@
 import { AptosAccount } from "aptos";
 import { DU } from "../src";
-import { PoolClient } from "../src/amm/pool";
+import { PoolClient } from "../src/pool/client";
 import { AuxClient } from "../src/client";
 import { FakeCoin } from "../src/coin";
 import { AuxEnv } from "../src/env";
